Renaming and refactoring stats files
