+++ webserver/views/stats.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 from flask import Blueprint, Response, render_template
-from db.stats import get_statistics_history
+from db.submission_stats import get_statistics_history
 from operator import itemgetter
 import json
